requirements.txt
cryptography==38.0.3
firebase-admin==6.0.1
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-logging==3.3.1
google-cloud-storage==2.6.0
PyJWT==2.6.0
requests==2.28.1
rsa==4.9
selenium==4.6.0
tqdm==4.64.1
webdriver-manager==3.8.5
